Allow testing a single file
Update the file path argument handling to allow testing a single file.

Error checking during config generation is removed, as we have to check
for missing file/directory errors anyway while running the tests, and
checking the files ahead of time isn't that useful as the file or
directory can disappear between the config generation and running the
tests.